requirements.txt
torch>=2.0.0
torchvision>=0.15.0
wandb>=0.15.0
pyyaml>=6.0
tqdm>=4.65.0
numpy>=1.24.0
pillow>=9.5.0